Data producer creation working (data producer itself is not), minor fixes

+++ rust/src/router/producer.rs
@@ -113,7 +113,7 @@ pub struct ProducerVideoOrientation {
 pub struct ProducerStat {
     // Common to all RtpStreams.
     pub r#type: RtpType,
-    pub timestamp: u32,
+    pub timestamp: u64,
     pub ssrc: u32,
     pub rtx_ssrc: Option<u32>,
     pub rid: Option<String>,
